Removing the negative
I think it's a bad idea to even bring up the fact that this is "a bit
of a cheat". We shouldn't say things like "this isn't some cheap and
dirty fix". No one has said anything about it being that, so you're
actually putting that in the readers mind. You may want to re-add some
things removed but I don't really think it's necessary.